1st commit of this project with utils file
Added an utils file to the project to get a few more details like: processor speed, size of the file that is being processed and a function to save this information in a csv file to create some benchmarking graph. The idea of this project is to participate of the 1 billion rows challenge and to add my touch on it, in this case I'm going to benchmark time to process the data on different languages/tools and different file sizes using different cpu clock speeds (configured through windows energy control).